requirements.txt
Flask==1.1.1
Flask-JWT-Extended==3.22.0
Flask-Babel==1.0.0
flask-marshmallow==0.10.1
neomodel==3.3.2
Flask-Cors==3.0.8
sentry-sdk==0.10.2
blinker==1.4
uwsgi==2.0.18
